inital pass at tablecard
